requirements.txt
Django==1.8
psycopg2==2.6.1
djangorestframework==3.1.1
django-rest-swagger==0.2.9

